HIP 15 | David | Refactoring the OpenMRS related discovery repository and move to separate folder
