--- src/In.ProjectEKA.HipService/Discovery/IPatientDal.cs
@@ -1,11 +1,0 @@
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using Hl7.Fhir.Model;
-
-namespace In.ProjectEKA.HipService.Discovery
-{
-    public interface IPatientDal
-    {
-        Task<List<Patient>> LoadPatientsAsync(string name, AdministrativeGender? gender, string yearOfBirth);
-    }
-}
