+++ src/In.ProjectEKA.HipService/In.ProjectEKA.HipService.csproj
@@ -61,4 +61,7 @@
         <ProjectReference Include="..\In.ProjectEKA.HipLibrary\In.ProjectEKA.HipLibrary.csproj" />
     </ItemGroup>
     
+    <ItemGroup>
+      <Folder Include="OpenMrs\" />
+    </ItemGroup>
 </Project>
