+++ test/In.ProjectEKA.HipServiceTest/Discovery/OpenMrsPatientMatchingRepositoryTest.cs
@@ -2,7 +2,6 @@ namespace In.ProjectEKA.HipServiceTest.Discovery
 {
     using System;
     using FluentAssertions;
-    using HipService.Discovery;
     using HipLibrary.Patient.Model;
     using Xunit;
     using Moq;
@@ -12,6 +11,7 @@ namespace In.ProjectEKA.HipServiceTest.Discovery
     using OpenMrsPatient = Hl7.Fhir.Model.Patient;
     using OpenMrsPatientName = Hl7.Fhir.Model.HumanName;
     using OpenMrsGender = Hl7.Fhir.Model.AdministrativeGender;
+    using In.ProjectEKA.HipService.OpenMrs;
 
     public class OpenMrsPatientMatchingRepositoryTest
     {
